backend/requirements.txt
pydantic
pydantic-settings
python-dotenv
motor
pymongo
# bson
fastapi 
uvicorn 
passlib[bcrypt] 
python-jose
httpx 
google-auth
requests
google-api-python-client